【準備】trick_tagにint Timeを追加　Actor.Spdをdoubleにして読み込みにも対応

+++ Battle/EnemySpeciesManager.cpp
@@ -32,7 +32,7 @@ void CEnemySpeciesManager::Clear(){
 
 }
 
-bool CEnemySpeciesManager::CreateSpecies(const char* _name, int _maxhp, int _atk, int _def, int _spd, int _img){
+bool CEnemySpeciesManager::CreateSpecies(const char* _name, int _maxhp, int _atk, int _def, double _spd, int _img){
 	CEnemySpecies newEnemy;
 	newEnemy.SetValue(_name, _maxhp, _atk, _def, _spd);
 	newEnemy.Img = _img;
